LPS-113018 Prevent split package by moving everything to layout-impl
